Kgj/chore/release0.2

+++ src/main/java/com/example/monthlylifebackend/subscribe/repository/RepairRequestRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepairRequestRepository extends JpaRepository<RepairRequest, Integer> {
+public interface RepairRequestRepository extends JpaRepository<RepairRequest, Long> {
 
 }
